feat(layout): custom Link component
This component has two possible styles: button and link. It was inspired
by the example provided by MUI
(https://github.com/mui/material-ui/blob/HEAD/examples/nextjs-with-typescript/src/Link.tsx).

TypeScript shoots out some errors about the ref and properties not being
compatible. This is due to the union of the MuiLinkProps and
MuiButtonProps. Even it being a tagged union does not seem to make a
different.
I couldn't figure out how to solve these errors, so I just decided to
make TypeScript ignore these lines. It's not the best solution, but
maybe in the future I can figure out a way to fix these errors.
For now, everything seems to work fine and specific props are recognized
beloging to the correct element.