feat(ui): add chakra library
After the adventure down to MUI energy-sucking types, I've decided it's
better to switch to another component library. When making edits I had to
wait upwards of 5-15secs for the TypeScript errors to update. This made
development impossible and completely unefficient.

Chakra seems to be a good and popular library, with good amount of
components.